just missing setValue for VariableBinding
