+++ package.json
@@ -31,6 +31,7 @@
     "mocha": "^8.4.0",
     "ts-mocha": "^8.0.0",
     "ts-node": "^10.1.0",
+    "tslog": "^3.2.0",
     "typescript": "^4.3.5"
   }
 }
