+++ src/Binding.ts
@@ -1,10 +1,13 @@
 import { Context, Property } from './UidType';
+import { Logger } from "tslog";
 
 export class Binding {
     property: Property;
     context: Context;
+    logger: Logger;
 
     constructor(property: Property, context: Context) {
+        this.logger = new Logger({ name: "Bidning" });
         this.property = property;
         this.context = context;
     }
@@ -12,7 +15,7 @@ export class Binding {
     setValue(value: string): void {
         throw new Error("Method not implemented.");
     }
-    getValue(): string {
+    getValue(): string | undefined {
         throw new Error("Method not implemented.");
     }
 }
@@ -23,75 +26,6 @@ export class ConstantBinding extends Binding {
     }
 }
 
-export class ExpressionBinding extends Binding {
-    getter: Function;
-    currentValue: string = '';
-
-    constructor(property: Property, context: Context) {
-        super(property, context);
-        this.getter = () => wrappedEval(`return ${property.value}`, context);
-    }
-
-    getValue() {
-        const newValue = this.getter(this.context, this.property.value);
-        if (!Object.is(this.currentValue, newValue)) {
-            this.currentValue = newValue;
-        }
-        return this.currentValue;
-    }
-}
-
-//TODO BP: Use translation mecanisme here
-export class InterpolationBinding extends Binding {
-
-    constructor(property: Property, context: Context) {
-        super(property, context);
-    }
-
-    getValue() {
-        console.warn('To be implemented - InterpolationBinding is not ready to production');
-        //Todo: make possible to use | from angular (ex: | uppercase, | date ...)
-        // return gettextCatalog.getString(this.property.value || '', this.context);
-
-        let finalValue = this.property.value;
-        // Match in value each {{ variable }} to replace it by real value
-
-
-        let arrStr = this.property.value.match(/({{.*?}})/gs) || [];
-        arrStr.forEach(match => {
-            let varName = match.slice(2, match.length - 2);
-            finalValue = this.property.value.replace(match, this.context[varName].displayValue);
-        });
-
-        return finalValue;
-    }
-}
-
-
-export class VariableBinding extends Binding {
-    getter: Function;
-    isBound: boolean;
-    value: String = '';
-
-    constructor(property: Property, context: Context) {
-        super(property, context);
-        this.value = property.value;
-        this.getter = () => wrappedEval(eval('return ${`' + property.value + '}`'), context);
-        this.isBound = !property.value;
-    }
-
-    getValue() {
-        return (!this.isBound) ? this.getter(this.context) : this.value;
-    }
-
-    setValue(newValue: string) {
-        const variableKey: string = Object.keys(this.context).filter(e => e == this.property.value)[0];
-        this.context[variableKey] = newValue;
-
-        return (!this.isBound) ? this.getter(this.context) : (this.value = newValue);
-    }
-}
-
 export enum enumBinding {
     constant = "constant",
     interpolation = "interpolation",
@@ -99,8 +33,16 @@ export enum enumBinding {
     expression = "expression"
 }
 
-
-function wrappedEval(expressionToEvaluate: string, contextData: any) {
-    // return (new Function(...Object.keys(contextData), expressionToEvaluate))(...Object.values(contextData));
-    return (new Function(`with(this) { ${expressionToEvaluate} }`)).call(contextData);
+/**
+ * Evaluate a js expression with context
+ * @param expressionToEvaluate 
+ * @param contextData Equivalent of scope
+ * @returns String | undefined if expression throw an error
+ */
+export function wrappedEval(expressionToEvaluate: string, contextData: any) {
+    try {
+        return (new Function(`with(this) { ${expressionToEvaluate} }`)).call(contextData);
+    } catch (e) {
+        return undefined;
+    }
 }
